added server and client framework base classes, added support for vector graphic fonts, and added multi-threading support to bullet physics

+++ SerenityLib/Engine/src/core/engine/mesh/VertexDataFormat.cpp
@@ -25,7 +25,7 @@ void VertexDataFormat::bind(const VertexData& vertData) {
 }
 void VertexDataFormat::unbind() { 
     for (size_t i = 0; i < attributes.size(); ++i) {
-        glDisableVertexAttribArray(i);
+        glDisableVertexAttribArray(static_cast<GLuint>(i));
     }
 }
